Implement application state management
Added a struct, StateInfo, which for now includes only a bool to check if the app is running. Passing StateInfo data into the CreateWindow function allows us to retrieve it in the message callback, eliminating the need for global variables to move data between the two.

We can expand StateInfo as need to allow for more versatility between the main program and the Windows message system.